+++ src/Oro/Bundle/CheckoutBundle/DependencyInjection/Compiler/CheckoutLineItemConverterPass.php
@@ -1,0 +1,23 @@
+<?php
+
+namespace Oro\Bundle\CheckoutBundle\DependencyInjection\Compiler;
+
+use Oro\Component\DependencyInjection\Compiler\TaggedServicesCompilerPassTrait;
+use Symfony\Component\DependencyInjection\Compiler\CompilerPassInterface;
+use Symfony\Component\DependencyInjection\ContainerBuilder;
+
+class CheckoutLineItemConverterPass implements CompilerPassInterface
+{
+    use TaggedServicesCompilerPassTrait;
+
+    const EXTENSION_TAG = 'oro.checkout.line_item.converter';
+    const SERVICE_ID = 'oro_checkout.line_item.converter_registry';
+
+    /**
+     * {@inheritDoc}
+     */
+    public function process(ContainerBuilder $container)
+    {
+        $this->registerTaggedServices($container, self::SERVICE_ID, self::EXTENSION_TAG, 'addConverter');
+    }
+}