+++ src/Oro/Bundle/CheckoutBundle/Migrations/Schema/OroCheckoutBundleInstaller.php
@@ -3,7 +3,6 @@
 namespace Oro\Bundle\CheckoutBundle\Migrations\Schema;
 
 use Doctrine\DBAL\Schema\Schema;
-
 use Oro\Bundle\MigrationBundle\Migration\Installation;
 use Oro\Bundle\MigrationBundle\Migration\QueryBag;
 
@@ -18,7 +17,7 @@ class OroCheckoutBundleInstaller implements Installation
      */
     public function getMigrationVersion()
     {
-        return 'v1_6';
+        return 'v1_7';
     }
 
     /**
@@ -30,9 +29,13 @@ class OroCheckoutBundleInstaller implements Installation
         $this->createOroCheckoutSourceTable($schema);
         $this->createOroCheckoutTable($schema);
         $this->createCheckoutWorkflowStateTable($schema);
+        $this->createOroCheckoutLineItemTable($schema);
+        $this->createOroCheckoutSubtotalTable($schema);
 
         /** Foreign keys generation **/
         $this->addOroCheckoutForeignKeys($schema);
+        $this->addOroCheckoutLineItemForeignKeys($schema);
+        $this->addOroCheckoutSubtotalForeignKeys($schema);
     }
 
     /**
@@ -166,5 +169,106 @@ class OroCheckoutBundleInstaller implements Installation
         $table->addColumn('updated_at', 'datetime', ['comment' => '(DC2Type:datetime)']);
         $table->addUniqueIndex(['entity_id', 'entity_class', 'token'], 'oro_checkout_wf_state_uidx');
         $table->setPrimaryKey(['id']);
+    }
+
+    /**
+     * Create oro_checkout_subtotal table
+     *
+     * @param Schema $schema
+     */
+    protected function createOroCheckoutSubtotalTable(Schema $schema)
+    {
+        $table = $schema->createTable('oro_checkout_subtotal');
+        $table->addColumn('id', 'integer', ['autoincrement' => true]);
+        $table->addColumn('checkout_id', 'integer', []);
+        $table->addColumn('currency', 'string', ['length' => 255]);
+        $table->addColumn(
+            'value',
+            'money',
+            ['notnull' => false, 'precision' => 19, 'scale' => 4, 'comment' => '(DC2Type:money)']
+        );
+        $table->addColumn('is_valid', 'boolean', []);
+        $table->addUniqueIndex(['checkout_id', 'currency'], 'unique_checkout_currency');
+        $table->setPrimaryKey(['id']);
+    }
+
+    /**
+     * Create oro_checkout_line_item table
+     *
+     * @param Schema $schema
+     */
+    protected function createOroCheckoutLineItemTable(Schema $schema)
+    {
+        $table = $schema->createTable('oro_checkout_line_item');
+        $table->addColumn('id', 'integer', ['autoincrement' => true]);
+        $table->addColumn('checkout_id', 'integer', ['notnull' => false]);
+        $table->addColumn('product_id', 'integer', ['notnull' => false]);
+        $table->addColumn('parent_product_id', 'integer', ['notnull' => false]);
+        $table->addColumn('product_unit_id', 'string', ['notnull' => false, 'length' => 255]);
+        $table->addColumn('product_sku', 'string', ['notnull' => false, 'length' => 255]);
+        $table->addColumn('free_form_product', 'string', ['notnull' => false, 'length' => 255]);
+        $table->addColumn('quantity', 'float', ['notnull' => false]);
+        $table->addColumn('product_unit_code', 'string', ['notnull' => false, 'length' => 255]);
+        $table->addColumn(
+            'value',
+            'money',
+            ['notnull' => false, 'precision' => 19, 'scale' => 4, 'comment' => '(DC2Type:money)']
+        );
+        $table->addColumn('currency', 'string', ['notnull' => false, 'length' => 255]);
+        $table->addColumn('price_type', 'integer', []);
+        $table->addColumn('from_external_source', 'boolean', []);
+        $table->addColumn('comment', 'text', ['notnull' => false]);
+        $table->addColumn('is_price_fixed', 'boolean', []);
+        $table->setPrimaryKey(['id']);
+    }
+
+    /**
+     * Add oro_checkout_line_item foreign keys.
+     *
+     * @param Schema $schema
+     */
+    protected function addOroCheckoutLineItemForeignKeys(Schema $schema)
+    {
+        $table = $schema->getTable('oro_checkout_line_item');
+        $table->addForeignKeyConstraint(
+            $schema->getTable('oro_checkout'),
+            ['checkout_id'],
+            ['id'],
+            ['onUpdate' => null, 'onDelete' => 'CASCADE']
+        );
+        $table->addForeignKeyConstraint(
+            $schema->getTable('oro_product'),
+            ['product_id'],
+            ['id'],
+            ['onUpdate' => null, 'onDelete' => 'CASCADE']
+        );
+        $table->addForeignKeyConstraint(
+            $schema->getTable('oro_product'),
+            ['parent_product_id'],
+            ['id'],
+            ['onUpdate' => null, 'onDelete' => 'CASCADE']
+        );
+        $table->addForeignKeyConstraint(
+            $schema->getTable('oro_product_unit'),
+            ['product_unit_id'],
+            ['code'],
+            ['onUpdate' => null, 'onDelete' => 'SET NULL']
+        );
+    }
+
+    /**
+     * Add oro_checkout_subtotal foreign keys.
+     *
+     * @param Schema $schema
+     */
+    protected function addOroCheckoutSubtotalForeignKeys(Schema $schema)
+    {
+        $table = $schema->getTable('oro_checkout_subtotal');
+        $table->addForeignKeyConstraint(
+            $schema->getTable('oro_checkout'),
+            ['checkout_id'],
+            ['id'],
+            ['onUpdate' => null, 'onDelete' => 'CASCADE']
+        );
     }
 }